requirements.txt
pandas>=1.5.0
numpy>=1.25.0
scikit-learn>=1.3.0


requests>=2.31.0
python-dotenv>=1.0.0


torch>=2.1.0
torchvision>=0.16.0


matplotlib>=3.8.0
seaborn>=0.12.2
